pyproject.toml
[project]
name = "researchos"
version = "0.1.0"
description = "ResearchOS research pipeline backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "apscheduler>=3.10.4",
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.4",
  "mcp>=1.13.1",
  "paramiko>=3.5.1",
  "pydantic-settings>=2.10.1",
  "pywinpty>=3.0.3; platform_system == 'Windows'",
  "python-dotenv>=1.1.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.43",
  "uvicorn>=0.35.0",
  "websockets>=16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "ruff>=0.12.12",
]
llm = [
  "openai>=1.102.0",
  "anthropic>=0.62.0",
]
pdf = [
  "pymupdf==1.26.7",
  "Pillow>=10.4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "packages*"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
  ".codex",
  ".venv",
  "backups",
  "data",
  "reference",
  "tmp",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [
  ".codex",
  ".git",
  ".venv",
  "backups",
  "data",
  "frontend",
  "reference",
  "tmp",
]
